mm: add support for remote mapping
The following two new mm exports are introduced:
 * mm_remote_map(struct mm_struct *req_mm,
                 unsigned long req_hva,
		 unsigned long map_hva)
 * mm_remote_unmap(unsigned long map_hva)

This patch allows one process to map into its address space a page from
another process. The previous page (if it exists) is dropped. There is
no corresponding pair of system calls as this API is meant to be used
by the kernel itself only.

The targeted user is the upcoming KVM VM introspection subsystem (KVMI),
where an introspector running in its own VM will map pages from the
introspected guest in order to eliminate round trips to the host kernel
(read/write guest pages).

The flow is as follows: the introspector identifies a guest physical
address where some information of interest is located. It creates a one
page anonymous mapping with MAP_LOCKED | MAP_POPULATE and calls the
kernel via an IOCTL on /dev/kvmmem giving the map virtual address and
the guest physical address as arguments. The kernel converts the map
va into a physical page (gpa in KVM-speak) and passes it to the host
kernel via a hypercall, along with the introspected guest gpa. The host
kernel converts the two gpa-s into their appropriate hva-s (host virtual
addresses) and makes sure the vma backing up the page belonging to the VM
in which the introspector runs, points to the indicated page into the
introspected guest. I have not included here the use of the mapping
token described in the KVMI documentation.

An introspector tool will always keep the most used mappings around
(tested with 4096).

There are some restrictions here:
 * the anonymous mapping created by the introspector must not change
   its backing physical page;
 * the page that ends up being shared must be locked in memory (ie.
   it cannot be paged out);
 * the mapping created adds to the mappings belonging to the
   controlling KVM process (qemu) corresponding to the VM inside which
   the introspector runs. /proc/sys/vm/max_map_count must be adjusted
   accordingly (eg. 16k mappings x 100 VMs = 1.6 million).

The following types might need some explaining, in order to ease reading
the patch:
 * struct intro_db - there is one such structure per each VM in which an
   introspector process is running. All instances are placed in a global
   hash table in order to avoid using space in mm_struct;
 * struct target_db - there is one such structure per each introspected
   VM. An intro_db instance can point to multiple target_db instances;
 * struct page_db - there is one such structure for each mapping. Thus,
   a target_db instance can point to multiple page_db instances.

TODO:
 * remove the module support and turn it into an optional built-in.
   This will avoid the need for exporting the guts of the memory manager;
 * figure out why this cannot be used with KSM enabled;
 * drop the intro/introspector naming;
 * test with more than one introspector (ie. multiple struct intro_db).

Signed-off-by: Mircea Cîrjaliu <mcirjaliu@bitdefender.com>