kvm: x86: hook in the VM introspection subsystem
Notify KVMI on vCPU create/destroy, VM destroy events, the new ioctl-s
(KVM_INTROSPECTION_HOOK, KVM_INTROSPECTION_UNHOOK) and when a KVMI
command is received (KVM_REQ_INTROSPECTION).

Change the x86 emulator and mmu behavior when KVMI is present (TODO:
do it when active) or when dealing with a page of interest for the
introspection tool.

Also, the EPT AD bits feature is disabled by this patch.

Signed-off-by: Adalbert Lazăr <alazar@bitdefender.com>
Signed-off-by: Mihai Donțu <mdontu@bitdefender.com>
Signed-off-by: Nicușor Cîțu <ncitu@bitdefender.com>
Signed-off-by: Mircea Cîrjaliu <mcirjaliu@bitdefender.com>