kvm: x86: hook in kvmi_trap_event()
Inform the guest introspection tool that the exception (from a previous
KVMI_INJECT_EXCEPTION command) was not successfully injected.

It can happen for the introspection tool to queue a pagefault but have
it overwritten by an interrupt picked up during guest reentry.
kvmi_trap_event() is used to inform the tool when such a situation has
been detected.

Signed-off-by: Mihai Donțu <mdontu@bitdefender.com>