KVM: introspection: add KVMI_CONTROL_CMD_RESPONSE
By default, all commands need a reply. This command allows to disable
replies for three classes:
    - replies for successful commands without data (eg. KVMI_VCPU_SET_REGISTERS)
    - replies for successful commands with data (eg. KVMI_VCPU_GET_REGISTERS)
    - replies for failed commands

It is useful when the introspection tool send multiple messages with
one write() call and doesn't have to wait for a reply.

IIRC, the speed improvment seen during UnixBench tests in a VM
introspected through vsock (the introspection tool was running in a
different VM) was around 5-10%.

Signed-off-by: Adalbert Lazăr <alazar@bitdefender.com>