kvm/mmu.c: use page tracking for GPAs with no access rights
The introspector tool might restrict all rights (RWX = 000) to some
GPAs, hence to some sptes. When an EPT violation is triggered by such a
page (the memory manager within the guest OS might access it), even
though it is marked as non-present because of the rights (RWX = 000),
the page still needs to be checked if it is tracked. Remove
PFERR_PRESENT_MASK check in order to achieve this.

Signed-off-by: Stefan Sicleru <ssicleru@bitdefender.com>