libkvmi: commit queued messages transparently when the batch limit is reached
The kvmi_batch_*() functions allow a user to send multiple commands
with one write() call.  Internally, these functions queue commands (when
the user calls kvmi_queue_*()) to be commited later with kvmi_batch_commit()
and the error handling for this function is complicated.

This commit changes kvmi_batch_add() to transparently commit the previous
messages if the current one doesn't have enough space.

A batch has two forms (use cases):

1)
        cmd 1 (disable command reply)
        cmd from user 1
        cmd from user 2
        ...                <- transparent commit
        cmd from user N-1
        cmd N (enable command reply and wait for a reply for this command)

2)
        cmd 1 (disable command reply)
        cmd from user 1
        cmd from user 2
        ...                <- transparent commit
        event reply
        cmd N (enable command reply without waiting for a reply for this command)

A transparent commit will:
  - flush the messages, appending
      "enable command reply without waiting for a reply for this command"
  - reset the batch, which will 'append' the "disable command reply"
      command for the next batch

It should be OK if another commands sneaks in (from a second thread)
when a batch is split.

Note: an event reply calls kvmi_batch_add() twice, once for the message
header and once for the payload. The transparent commit won't happend
between these two.

Signed-off-by: Adalbert Lazăr <alazar@bitdefender.com>